jekyll: ignore `*.lhs` files
When writing Literate Haskell posts, said files need the `lhs`
extension, whilst Jekyll requires posts to have the `md` extension to
render them as Markdown.

To avoid duplicates, posts are written as `lhs` files, and symlinked
with the `md` extension. The `lhs` files are now ignored, so there are
no duplicates.

Published posts are `md` only.